# ArmSwashbuckleStarterKit/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Microsoft.ArmSwashbuckleStarterKit",
    "description": "An Azure Resource Manager resource provider providing weather forecasts",
    "contact": {
      "name": "Jim Flowers",
      "url": "https://eng.ms/weather",
      "email": "scatteredshowers@microsoft.com"
    },
    "version": "2021-08-11-preview"
  },
  "paths": {
    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Weather/forecasts/{forecastId}": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "summary": "Fetches a weather forecast",
        "operationId": "WeatherForecast_Get",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "../../../../../common-types/resource-management/v3/types.json#/parameters/SubscriptionIdParameter"
          },
          {
            "$ref": "../../../../../common-types/resource-management/v3/types.json#/parameters/ResourceGroupNameParameter"
          },
          {
            "in": "path",
            "name": "forecastId",
            "required": true,
            "type": "string"
          },
          {
            "$ref": "../../../../../common-types/resource-management/v3/types.json#/parameters/ApiVersionParameter"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/WeatherForecast"
            }
          },
          "default": {
            "description": "Error response describing why the operation failed.",
            "x-ms-error-response": true,
            "schema": {
              "$ref": "../../../../../common-types/resource-management/v3/types.json#/definitions/ErrorResponse"
            }
          }
        },
        "x-ms-examples": {
          "Get": {
            "$ref": "./examples/Get.json"
          }
        }
      }
    },
    "/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Weather/forecasts/{forecastId}/fiveDay": {
      "get": {
        "tags": [
          "WeatherForecast"
        ],
        "summary": "Fetches a forecast and the next 4 days' forecasts",
        "operationId": "WeatherForecast_List",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "../../../../../common-types/resource-management/v3/types.json#/parameters/SubscriptionIdParameter"
          },
          {
            "$ref": "../../../../../common-types/resource-management/v3/types.json#/parameters/ResourceGroupNameParameter"
          },
          {
            "in": "path",
            "name": "forecastId",
            "required": true,
            "type": "string"
          },
          {
            "$ref": "../../../../../common-types/resource-management/v3/types.json#/parameters/ApiVersionParameter"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/WeatherForecast"
            }
          },
          "default": {
            "description": "Error response describing why the operation failed.",
            "x-ms-error-response": true,
            "schema": {
              "$ref": "../../../../../common-types/resource-management/v3/types.json#/definitions/ErrorResponse"
            }
          }
        },
        "x-ms-pageable": {
          "nextLinkName": "nextLink"
        },
        "x-ms-examples": {
          "List": {
            "$ref": "./examples/List.json"
          }
        }
      }
    }
  },
  "definitions": {
    "WeatherForecast": {
      "type": "object",
      "allOf": [
        {
          "$ref": "../../../../../common-types/resource-management/v3/types.json#/definitions/TrackedResource"
        }
      ],
      "properties": {
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "temperatureC": {
          "format": "int32",
          "type": "integer"
        },
        "temperatureF": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "summary": {
          "type": "string"
        },
        "publicId": {
          "type": "string"
        },
        "systemData": {
          "$ref": "../../../../../common-types/resource-management/v3/types.json#/definitions/systemData"
        }
      },
      "x-ms-azure-resource": true
    }
  },
  "securityDefinitions": {
    "azure_auth": {
      "type": "oauth2",
      "flow": "implicit",
      "authorizationUrl": "https://login.microsoftonline.com/common/oauth2/authorize",
      "scopes": {
        "user_impersonation": "impersonate your user account"
      },
      "description": "Azure Active Directory OAuth2 Flow"
    }
  }
}